Extract source location
